refactor DocumentURIString to .types
